Reorganize the genes onboarding step component
