+++ src/v2/Components/Onboarding/GenesStep/index.tsx
@@ -1,0 +1,1 @@
+export * from "./GenesStep"
